Move command-palette from components to features
