+++ src/features/command-palette/constants/advanced-actions.tsx
@@ -1,6 +1,6 @@
 import { invoke } from "@tauri-apps/api/core";
 import { RefreshCw, Sparkles, Terminal } from "lucide-react";
-import type { Action } from "../../models/action.types";
+import type { Action } from "../models/action.types";
 
 interface AdvancedActionsParams {
   lspStatus: {
